(1) sample lengths in fw are no longer confusingly in units of bytes (2) digital potentiometer code no longer has confusing byte swapping trick (3) switched to doubles in sw for samples (4) 0.018 ohm resistor created too small for amplifier, so had to be physicall switched with 0.18 ohm (and in params.h too)

+++ fw/samples.h
@@ -6,11 +6,11 @@
 extern uint32_t g_samples_ovsamp_bits;
 extern uint32_t g_samples_uart_seqnum;
 extern uint16_t g_samples_read_file;
-extern uint16_t g_adca0[], g_adca1[], g_adcb0[], g_adcb1[];
+extern int16_t g_adca0[], g_adca1[], g_adcb0[], g_adcb1[];
 
 uint16_t samples_ovsample(int16_t* v_s, int16_t* i_s, uint16_t len);
-void samples_uart_write(uint8_t* v_s, uint8_t* i_s, uint16_t len);
-void samples_store_write(uint8_t* v_s, uint8_t* i_s);
+void samples_uart_write(int16_t* v_s, int16_t* i_s, uint16_t len);
+void samples_store_write(int16_t* v_s, int16_t* i_s);
 void samples_store_read(uint16_t file);
 
 #endif
